# documentation/api.yaml
...
swagger: "2.0"
info:
  description: ""
  version: "0.0.5"
  title: "Urban Trees API"
  termsOfService: ""
  contact:
    email: "laurenz.fiala@outlook.at"
  license:
    name: ""
    url: "test://test"
host: "test.test"
basePath: "/test"
...
schemes:
- "https"

#
# Paths
#
paths:
  /tree:
    get:
      tags:
      - "tree"
      summary: "Return information about all trees"
      description: ""
      produces:
      - "application/json"
      responses:
        200:
          description: "Successful"
          schema:
            type: array
            items:
              $ref: "#/definitions/Tree"
        400:
          description: "Invalid tree id"
      security:
      - api_key: []
  /tree/{treeId}:
    get:
      tags:
      - "tree"
      summary: "Return information about a single tree"
      description: ""
      produces:
      - "application/json"
      parameters:
      - name: "treeId"
        in: "path"
        description: "ID of a tree"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Successful"
          schema:
            $ref: "#/definitions/Tree"
        400:
          description: "Invalid tree id"
      security:
      - api_key: []
...
securityDefinitions:
  api_key:
    type: apiKey
    name: x-api-key
    in: header

#
# Definitions
#
definitions:
...
  Tree:
    type: "object"
    required:
      - "id"
      - "location"
      - "speciesId"
      - "species"
      - "genusId"
      - "genus"
    properties:
      id:
        type: "integer"
        format: "int32"
      location:
        $ref: "#/definitions/TreeLocation"
      speciesId:
        type: "integer"
        format: "int32"
      species:
        type: "string"
      genusId:
        type: "integer"
        format: "int32"
      genus:
        type: "string"
      plantationYear:
        description: "Latest observed plantation year"
        type: "integer"
        format: "int32"
      isPlantationYearEstimate:
        description: "Whether lastest observed plantationYear is an estimate or not"
        type: "boolean"
...
  TreeLocation:
    type: "object"
    required:
      - "id"
    properties:
      id:
        type: "integer"
        format: "int32"
      coordinates:
        $ref: "#/definitions/Coordinates"
      street:
        type: "string"
      city:
        type: "string"
  Coordinates:
    type: "object"
    required:
      - "x"
      - "y"
    properties:
      x:
        type: "number"
        format: "float"
      y:
        type: "number"
        format: "float"
        
  # Error Responses
...